Added an empty output file for PHP errors.
Simply as that, added the log folder and decided to place a clean file
into it. I didnt put the Logs folder into /System, because this is more
relevant to the whole app. However, I will use that folder to log a lot
of things - like installations, patches and updates. So I might even
need a script to clean this folder if it gets too messy at times x)