Added a concept for a Installer-interface
This is simply a help for me so far, to know what goes where and
how…like a little booklet of notes and remarks. However - this can also
be used by others to port other methods such as APT to drag0n. Just a
fun idea, but maybe there are other crazy people like me out there :)