Security improvement - GnuPG is now installed into PHP!
This basically means, we have gpg+ssh2 - now we can fetch packages and
resource data and what not completely securily for all these people
that are a little for that stuff :). Now, back to the core. We will
support ssh2 and gpg at once from now on. a gpg binary and the gpgme
library may be compiled into the System folder too at some point for
all those that dont have it, but need it. :D