Merded Chromium into drag0n itself, introducing InfoSwap to swap Info.plist files depending on application
Yep, the hackery keeps going :)
Check Resources/script and see the class at the top. It works but
its…dirty… o-o But in fact, it works so good that I might even keep
that unless I can find a way to set the first PHP process to suicide.
The issue is that the /usr/bin/php process won't terminate, so the
following PHP process gets stuck in the icon too. The result are 2 dock
icons. The InfoSwap trick gets around that by telling the OS that there
are TWO applications in one bundle, essentialy making it treat it as
such. Phew…hackery man…hackery :p